+++ packages/vue/src/presence/index.ts
@@ -1,0 +1,8 @@
+import { Presence as PresenceRoot, type PresenceProps } from './presence'
+
+const Presence = Object.assign(PresenceRoot, {
+  Root: PresenceRoot,
+})
+
+export { Presence }
+export type { PresenceProps }