+++ packages/vue/src/tooltip/tooltip-arrow-tip.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useTooltipContext } from './tooltip-context'
 
 export type TooltipArrowTipProps = HTMLArkProps<'div'>
 
-export const TooltipArrowTip: ComponentWithProps<TooltipArrowTipProps> = defineComponent({
+export const TooltipArrowTip = defineComponent({
   name: 'TooltipArrowTip',
   setup(_, { slots, attrs }) {
     const api = useTooltipContext()