+++ packages/vue/src/tooltip/tooltip-context.ts
@@ -1,7 +1,6 @@
 import { createContext } from '../context'
 import { type UseTooltipReturn } from './use-tooltip'
 
-export const [TooltipProvider, useTooltipContext] =
-  createContext<UseTooltipReturn>('TooltipContext')
-
 export type TooltipContext = UseTooltipReturn
+
+export const [TooltipProvider, useTooltipContext] = createContext<TooltipContext>('TooltipContext')