+++ packages/vue/src/tooltip/tooltip-positioner.tsx
@@ -1,11 +1,11 @@
 import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useTooltipContext } from './tooltip-context'
 
 export type TooltipPositionerProps = HTMLArkProps<'div'>
 
-export const TooltipPositioner: ComponentWithProps<TooltipPositionerProps> = defineComponent({
+export const TooltipPositioner = defineComponent({
   name: 'TooltipPositioner',
   setup(_, { slots, attrs }) {
     const api = useTooltipContext()